% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-clc-coordinates.R
\name{extract_clc_coordinates}
\alias{extract_clc_coordinates}
\title{Extract start and end coordinates from a CLC genomic region string}
\usage{
extract_clc_coordinates(df, coord_col)
}
\arguments{
\item{df}{A dataframe containing a column of genomic regions}

\item{coord_col}{The region column}
}
\value{
The input dataframe with the region start and end coordinates
as two new columns.
}
\description{
Extract start and end coordinates from a CLC genomic region string
}
\examples{

df <- data.frame("region" = c("55174776..55174793"))

df |>
extract_clc_coordinates(coord_col = region)

}
