% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-coordinate.R
\name{extract_coordinate}
\alias{extract_coordinate}
\title{Extract a genomic coordinate from a character string}
\usage{
extract_coordinate(x, location = "start", type = "numeric")
}
\arguments{
\item{x}{A character string containing one or two genomic coordinates}

\item{location}{"start" or "end"}

\item{type}{The type of value to return: "numeric" or "character"}
}
\value{
The extracted genomic coordinate
}
\description{
Extract a genomic coordinate from a character string
}
\examples{
extract_coordinate("1234_12345", location = "end")
}
