% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorise-chromosome.R
\name{factorise_chromosome}
\alias{factorise_chromosome}
\title{Convert human chromosomes into factors}
\usage{
factorise_chromosome(x)
}
\arguments{
\item{x}{Character string or character vector}
}
\value{
The input formatted as a factor with consistent chromosome levels.
}
\description{
Human chromosomes contain 1-22 autosomes and X and Y sex chromosomes. This
combination of numbers and letters means that chromosomes are best formatted
as factors for data analysis.
\code{factorise_chromosome} takes a character or vector input of chromosome names
and converts them into factors to allow for easier analysis and plotting.
}
\examples{
df <- data.frame("chromosome" = c("1", "2", "10", "X")) |>
dplyr::arrange(chromosome)

print(df)

df |>
dplyr::mutate(chromosome = factorise_chromosome(chromosome)) |>
dplyr::arrange(chromosome)

}
