% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-chromosomes.R
\name{levels_chromosomes}
\alias{levels_chromosomes}
\title{Factor levels for human chromosomes}
\usage{
levels_chromosomes()
}
\value{
A list of character vectors of human chromosomes in order, with
options for autosomes, sex chromosomes and all chromosomes (with sex
chromosomes ordered after autosomes)
}
\description{
Factor levels for human chromosomes
}
\examples{

levels_chromosomes()$all

}
