% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-coordinates.R
\name{mutate_coordinates}
\alias{mutate_coordinates}
\title{Add genomic coordinates in separate columns of a dataframe}
\usage{
mutate_coordinates(df, coord_col)
}
\arguments{
\item{df}{A dataframe}

\item{coord_col}{The column containing genomic coordinate strings}
}
\value{
The dataframe with coordinates extracted into "start" and "end"
columns
}
\description{
Add genomic coordinates in separate columns of a dataframe
}
\examples{

test_df <- tibble::tribble(
~"region",
"1",
"1..3",
"123456789..234567890")

test_df |> mutate_coordinates(coord_col = region)

}
