% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-region.R
\name{regex_region}
\alias{regex_region}
\title{Regular expressions for genomic region coordinate strings}
\usage{
regex_region()
}
\value{
A named list of regular expressions and coordinate groupings
}
\description{
Descriptions of genomic regions are often given as character strings rather
than as separate columns of start and end coordinates.
\code{regex_region} contains regular expressions for describing different formats
of character string, to allow start and end coordinates to be easily
extracted.
}
\examples{
stringr::str_extract(string = "1234..5678",
pattern = regex_region()$clc$regex,
group = regex_region()$clc$start)

}
